fix extension auth
